Mades Changes to UI of Create Profile and Got Store Locator Feature Working
I made the UI of the create profile pages be consistent with the UI of the login page. I also got the frontend and backend features of the store locator feature to work so now on the species information page for a plant, the nearest plant stores based on the user's location are displayed. The user is also prompted to grant the web page location access.